Xcode warnings + Xcode project file mods (#1563)

+++ src/rtFileDownloader.cpp
@@ -689,7 +689,7 @@ void rtFileDownloader::downloadFile(rtFileDownloadRequest* downloadRequest)
             if(fp != NULL)
             {
                 char* buffer = new char[downloadRequest->getCachedFileReadSize()];
-                int bytesCount = 0;
+                size_t bytesCount = 0;
                 size_t dataSize = 0;                
 				char invalidData[8] = "Invalid";
